+++ lib/src/exceptions/th_create_object_without_list.dart
@@ -1,11 +1,9 @@
-class THCreateObjectWithoutListException implements Exception {
-  String objectType;
-  dynamic originalInfo;
+import 'package:mapiah/src/exceptions/th_base_exception.dart';
 
-  THCreateObjectWithoutListException(this.objectType, this.originalInfo);
+class THCreateObjectWithoutListException extends THBaseException {
+  final String objectType;
+  final dynamic originalInfo;
 
-  @override
-  String toString() {
-    return "Can´t create object of type '$objectType' from '$originalInfo.";
-  }
+  THCreateObjectWithoutListException(this.objectType, this.originalInfo)
+    : super("Can´t create object of type '$objectType' from '$originalInfo.");
 }